yololab/pyproject.toml
[project]
name = "yololab"
version = "0.1.0"
description = ""
authors = [
    {name = "Stefano Scipioni", email = "stefano.scipioni@csgalileo.net"},
]
dependencies = [
    "configargparse>=1.5.3",
    #"ultralytics>=8.0.115",
    #"shared-ndarray2>=1.0.0.post0",
    "ipython>=8.13.2",
    #"deep-sort-realtime>=1.3.2",
    #"super-gradients>=3.1.2",
    "pascal-voc>=0.0.7",
    "imagesize>=1.4.1",
    #"torchinfo>=1.8.0",
    #"onnx>=1.14.0",
    #"opencv-python-rolling<5.0.0",
    "shared-ndarray2>=1.0.0.post0",
]
requires-python = ">=3.9"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[project.optional-dependencies]  # This is optional dependencies
train = [
"ultralytics>=8.0.117",
"torchinfo>=1.8.0",
"onnx>=1.14.0",
"opencv-python-rolling<5.0.0",
]
#extra2 = ["django"]


[project.scripts]
grabber = "yololab.main:main"
fpds2voc = "yololab.utils.fpds:main"
dynamic-cropper = "yololab.utils.DatasetDynamicCropper.main:main"
dsconvert = "yololab.utils.dsconvert.main:main"
